// source/typings/deviceType.ts
export enum DeviceType {
    Cooking = "devices.types.cooking",
    CookingCoffee = "devices.types.cooking.coffee_maker",
    CookingKettle = "devices.types.cooking.kettle",
    CookingMulticooker = "devices.types.cooking.multicooker",

    Light = "devices.types.light",
    LightCeiling = "devices.types.light.ceiling",
    LightStrip = "devices.types.light.strip",

    Media = "devices.types.media_device",
    MediaReceiver = "devices.types.media_device.receiver",
    MediaTv = "devices.types.media_device.tv",
    MediaTvBox = "devices.types.media_device.tv_box",

    Openable = "devices.types.openable",
    OpenableCurtain = "devices.types.openable.curtain",
    OpenableValve = "devices.types.openable.valve",

    PetFountain = "devices.types.pet_drinking_fountain",
    PetFeeder = "devices.types.pet_feeder",

    Sensor = "devices.types.sensor",
    SensorButton = "devices.types.sensor.button",
    SensorClimate = "devices.types.sensor.climate",
    SensorGas = "devices.types.sensor.gas",
    SensorIllumination = "devices.types.sensor.illumination",
    SensorMotion = "devices.types.sensor.motion",
    SensorOpen = "devices.types.sensor.open",
    SensorSmoke = "devices.types.sensor.smoke",
    SensorVibration = "devices.types.sensor.vibration",
    SensorWaterLeak = "devices.types.sensor.water_leak",

    Meter = "devices.types.smart_meter",
    MeterColdWater = "devices.types.smart_meter.cold_water",
    MeterElectricity = "devices.types.smart_meter.electricity",
    MeterGas = "devices.types.smart_meter.gas",
    MeterHeat = "devices.types.smart_meter.heat",
    MeterHotWater = "devices.types.smart_meter.hot_water",

    Thermostat = "devices.types.thermostat",
    ThermostatAc = "devices.types.thermostat.ac",

    Ventilation = "devices.types.ventilation",
    VentilationFan = "devices.types.ventilation.fan",
    
    Camera = "devices.types.camera",
    Dishwasher = "devices.types.dishwasher",
    Humidifier = "devices.types.humidifier",
    Iron = "devices.types.iron",
    Other = "devices.types.other",
    Purifier = "devices.types.purifier",
    Socket = "devices.types.socket",
    Switch = "devices.types.switch",
    VacuumCleaner = "devices.types.vacuum_cleaner",
    WashingMachine = "devices.types.washing_machine"
}